requirements.txt
python-telegram-bot 
python-dotenv
openai
python-dotenv
sqlalchemy
